requirements.txt
numpy==1.11.0
python-Levenshtein==0.12.0
